Add some testing, rename script
